more godoc comments
